refactor(LinesGenerator): line 패키지 아래로 파일 위치 변경
