+++ src/test/java/model/LineTest.java
@@ -3,7 +3,7 @@ package model;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import model.path.Path;
+import model.line.Line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
