+++ src/test/java/model/line/FixedLinesGenerator.java
